Force Apps to be Pre-Processed Last
It seems that nanoc does not order files based off of name alone and the
order of files can change based on access time (or similarattribute
changes).

In the current state of this branch on a fresh installation (from RPM)
the apps.html.erb file attempts to process before the rest (which seems
correct due to alphabetical things) however this will cause a "binary?"
error for "nil:NilClass" (likely related to some of the nesting/looping
caused by the "compiled_content" calls).

I'm not sure that I'm explaining this issue correctly but this solution
works so let's try not to question it together.